Update TensorFlow Large Model Support examples
Update the TensorFlow Large Model Support examples to support TensorFlow 1.15 and 2.0.0.

Add the Keras_ManyModel example and base the Keras_ResNet50 example on it.
