+++ examples/tensorflow_large_model_support/v2/mnist_deep_lms.py
@@ -33,7 +33,8 @@ import tempfile
 
 from tensorflow.examples.tutorials.mnist import input_data
 
-import tensorflow as tf
+import tensorflow.compat.v1 as tf
+tf.disable_v2_behavior()
 
 import numpy
